scripts/requirements.txt
# Database
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Background Jobs
celery==5.3.4

# AI/ML
transformers==4.36.2
huggingface-hub==0.20.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.4
black==23.12.1
flake8==7.0.0